Added stopping rule and checked for different starting values.
Convergence is assumed when log likelihood of previous and current step
differ with a constant smaller than  0.001.

ADVICE FOR STARTING VALUES:
Should not take all the same starting values for the 3 parameters as the algorithm cannot handle the same
probabilities. Starting values that are far from true values do seem to converge to the
true values but take more iterations.