Updating figures, adding code to calculate summary measurements and changing to average overlapping clusters.
Bunch of code to update figures for the manuscript.
Added code to calculate various summaries on the measurmentes of interest (for results section).
Also switched to the average of non-overlapping versus overlapping clusters.
This average is calculated over the test and replication image as it makes more sense (one can
talk about the individual analyses instead of having to know that it is both images we are looking at).

Will remove all comments and clean up code that has been used for OHBM in following commit.