Extending adaptive thresholding and adding more baseline proportions.
First we now have baseline proportions of activation equal to
0, .05, .20, .80 and 1.

The target percentage is the amount of voxels aimed for in the adaptive
procedure.

For .20 and .80, we now look at target percentages of the baseline,
.10 below or .10 above the baseline.
For 0, we take .05 as the target.
For 1, we take .95 as the target.
For .05, we take .05 as the target.