Added second algorithm which uses a percentile approach and added estimation of m1.
Percentile approach takes the target percentage as the percentile value
on the sorted P-values.
Added this as an argument so you can choose between adaptive and percentile approach.
Furthermore, added the estimation of M1 on the functions obtained
by percentile and adaptive thresholding.
These are different functions and therefor I use different estimation approaches.